refactor(cli): split local control tests
